bump to v2.0.0+rev1

+++ src/main/java/com/smartx/tower/model/DatacenterWhereInput.java
@@ -381,7 +381,7 @@ public class DatacenterWhereInput {
 
   public static final String SERIALIZED_NAME_USED_CPU_HZ = "used_cpu_hz";
   @SerializedName(SERIALIZED_NAME_USED_CPU_HZ)
-  private Long usedCpuHz;
+  private Double usedCpuHz;
 
   public static final String SERIALIZED_NAME_USED_CPU_HZ_GT = "used_cpu_hz_gt";
   @SerializedName(SERIALIZED_NAME_USED_CPU_HZ_GT)
@@ -445,7 +445,7 @@ public class DatacenterWhereInput {
 
   public static final String SERIALIZED_NAME_USED_MEMORY_BYTES = "used_memory_bytes";
   @SerializedName(SERIALIZED_NAME_USED_MEMORY_BYTES)
-  private Long usedMemoryBytes;
+  private Double usedMemoryBytes;
 
   public static final String SERIALIZED_NAME_USED_MEMORY_BYTES_GT = "used_memory_bytes_gt";
   @SerializedName(SERIALIZED_NAME_USED_MEMORY_BYTES_GT)
@@ -2640,7 +2640,7 @@ public class DatacenterWhereInput {
   }
 
 
-  public DatacenterWhereInput usedCpuHz(Long usedCpuHz) {
+  public DatacenterWhereInput usedCpuHz(Double usedCpuHz) {
     
     this.usedCpuHz = usedCpuHz;
     return this;
@@ -2653,12 +2653,12 @@ public class DatacenterWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUsedCpuHz() {
+  public Double getUsedCpuHz() {
     return usedCpuHz;
   }
 
 
-  public void setUsedCpuHz(Long usedCpuHz) {
+  public void setUsedCpuHz(Double usedCpuHz) {
     this.usedCpuHz = usedCpuHz;
   }
 
@@ -3040,7 +3040,7 @@ public class DatacenterWhereInput {
   }
 
 
-  public DatacenterWhereInput usedMemoryBytes(Long usedMemoryBytes) {
+  public DatacenterWhereInput usedMemoryBytes(Double usedMemoryBytes) {
     
     this.usedMemoryBytes = usedMemoryBytes;
     return this;
@@ -3053,12 +3053,12 @@ public class DatacenterWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUsedMemoryBytes() {
+  public Double getUsedMemoryBytes() {
     return usedMemoryBytes;
   }
 
 
-  public void setUsedMemoryBytes(Long usedMemoryBytes) {
+  public void setUsedMemoryBytes(Double usedMemoryBytes) {
     this.usedMemoryBytes = usedMemoryBytes;
   }
